backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
supabase>=2.4.0
scikit-learn>=1.4.0
numpy>=1.26.0
pydantic>=2.7.0
twilio>=9.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
